requirements.txt
discord.py
yfinance
pandas
numpy
ta
python-dotenv
pytz